requirements.txt
pydantic-ai>=0.3
langfuse-python>=2.0
jinja2>=3.1
sqlmodel>=0.0.16
aiosqlite>=0.19
typer>=0.9
structlog>=24.1
rich>=13.7
pytest>=8.2
watchdog>=4.0
pyyaml>=6.0
